refactor: replace interface about us
