+++ src/components/AboutUs/AboutTabs/types/index.ts
@@ -1,0 +1,10 @@
+export interface IAboutDataTabs {
+  profilePicture: string;
+  stName: string;
+  role: string;
+  github: string;
+  body: {
+    textAbout: string;
+    recommendations: string[];
+  };
+}
